+++ branches/2.0/doubango/tinyNET/src/dhcp/tnet_dhcp.c
@@ -130,7 +130,7 @@ tnet_dhcp_reply_t* tnet_dhcp_send_request(tnet_dhcp_ctx_t* ctx, tnet_dhcp_reques
 			if (request->type == dhcp_type_inform){
 				struct sockaddr_storage ss;
 				if (!tnet_getsockname(localsocket4->fd, &ss)){
-					uint32_t addr = tnet_htonl_2(&((struct sockaddr_in*)&ss)->sin_addr);
+					uint32_t addr = (uint32_t)tnet_htonl_2(&((struct sockaddr_in*)&ss)->sin_addr);
 					memcpy(&request->ciaddr, &addr, 4);
 				}
 			}
